Поправил дребезг фильтров
