+++ js/main.js
@@ -1,9 +1,9 @@
-import {getPhotos} from './utils/server.js';
-import {showError, showPhotos} from './utils/show-photos.js';
-import {setPhotoListFilter} from './utils/photo-list-filter.js';
-import {setUploadButtonClick} from './utils/upload-new-foto.js';
-import {setUploadFormSubmit} from './utils/upload-form.js';
-import {createSlider} from './utils/effects.js';
+import {getPhotos} from './server.js';
+import {showError, showPhotos} from './show-photos.js';
+import {setPhotoListFilter} from './photo-list-filter.js';
+import {setUploadButtonClick} from './upload-new-foto.js';
+import {setUploadFormSubmit} from './upload-form.js';
+import {createSlider} from './effects.js';
 
 getPhotos((photos) => {
   showPhotos(photos);
