+++ js/scale-control.js
