+++ js/upload-new-foto.js
